refactor: Added vip number 181 to add cut param to diamond proxy
