+++ simulations/vip-181/vip-181/simulations.ts
@@ -6,22 +6,22 @@ import { ethers } from "hardhat";
 
 import { initMainnetUser } from "../../../src/utils";
 import { forking, pretendExecutingVip, testVip } from "../../../src/vip-framework";
-import { vipCutParamAdd } from "../../../vips/vip-diamond-cut-param-add/vip-param-add-testnet";
+import { vip181 } from "../../../vips/vip-181/vip-181";
 import Comptroller from "../abi/Comptroller.json";
 import IERC20Upgradeable from "../abi/IERC20UpgradableAbi.json";
 import VBEP20_DELEGATE_ABI from "../abi/VBep20DelegateAbi.json";
 import VENUS_LENS_ABI from "../abi/VenusLens.json";
 
-const UNITROLLER = "0x94d1820b2D1c7c7452A163983Dc888CEC546b77D";
-const VENUS_LENS = "0x36B434654bD5fb010f8A68e190428dc4789E1b24";
-const Owner = "0xce10739590001705F7FF231611ba4A48B2820327";
-const VBUSD = "0x08e0A5575De71037aE36AbfAfb516595fE68e5e4";
-const VUSDT = "0xb7526572FFE56AB9D7489838Bf2E18e3323b441A";
-const USER = "0xC825AD791A6046991e3706b6342970f6d87e4888";
+const UNITROLLER = "0xfD36E2c2a6789Db23113685031d7F16329158384";
+const VENUS_LENS = "0xfB0f09dB330dC842a6637BfB959209424BbFE8C7";
+const Owner = "0x939bD8d64c0A9583A7Dcea9933f7b21697ab6396";
+const VBUSD = "0x95c78222B3D6e262426483D42CfA53685A67Ab9D";
+const VUSDT = "0xfD5840Cd36d94D7229439859C0112a4185BC0255";
+const USER = "0x8894E0a0c962CB723c1976a4421c95949bE2D4E3";
 
 // Added function signature for venusInitialIndex to market facet as cut param to diamond proxy
 // This fork block contains tests for venusInitialIndex only.
-forking(33763885, async () => {
+forking(32159070, async () => {
   let diamondUnitroller: ethers.Contract;
   let venusLens: ethers.Contract;
 
@@ -40,7 +40,7 @@ forking(33763885, async () => {
     });
   });
 
-  testVip("VIP-Diamond cut param add", vipCutParamAdd());
+  testVip("VIP-Diamond cut param add", vip181());
 
   describe("After execution of vip", async () => {
     it("Fetching of VenusInitialIndex should return value", async () => {
@@ -58,7 +58,7 @@ forking(33763885, async () => {
 // As this vip is updating diamond proxy for the first time after it's implementation,
 // therefore adding additional tests to verify all storage values before and after executing this vip,
 // setter functions and core functionalities are working properly
-forking(33763885, async () => {
+forking(32159070, async () => {
   let owner: Signer,
     unitroller: Contract,
     // layout variables
@@ -112,7 +112,7 @@ forking(33763885, async () => {
       data: undefined,
     });
 
-    busdHolder = await initMainnetUser("0xC825AD791A6046991e3706b6342970f6d87e4888", parseUnits("1000", 18));
+    busdHolder = await initMainnetUser("0x8894E0a0c962CB723c1976a4421c95949bE2D4E3", parseUnits("1000", 18));
 
     [vBUSD, vUSDT] = await Promise.all(
       [VBUSD, VUSDT].map((address: string) => {
@@ -199,7 +199,7 @@ forking(33763885, async () => {
     });
   });
 
-  testVip("VIP-Diamond cut param add", vipCutParamAdd());
+  testVip("VIP-Diamond Contract Migration", vip181());
 
   describe("Verify Storage slots after VIP execution", async () => {
     // These tests checks the storage collision of comptroller while updating it via diamond.
@@ -360,7 +360,7 @@ forking(33763885, async () => {
   });
 });
 
-forking(33763885, async () => {
+forking(32159070, async () => {
   let owner, unitroller;
   let USDT: Contract;
   let usdtHolder: Signer;
@@ -368,7 +368,7 @@ forking(33763885, async () => {
   let diamondUnitroller: Contract;
 
   before(async () => {
-    await pretendExecutingVip(vipCutParamAdd());
+    await pretendExecutingVip(vip181());
     unitroller = new ethers.Contract(UNITROLLER, Comptroller, ethers.provider);
 
     diamondUnitroller = new ethers.Contract(unitroller.address, Comptroller, ethers.provider);
@@ -382,7 +382,7 @@ forking(33763885, async () => {
       data: undefined,
     });
 
-    usdtHolder = await initMainnetUser("0xa0747a72C329377C2CE4F0F3165197B3a5359EfE", parseUnits("1000", 18));
+    usdtHolder = await initMainnetUser("0x8894E0a0c962CB723c1976a4421c95949bE2D4E3", parseUnits("1000", 18));
 
     [vUSDT] = await Promise.all(
       [VUSDT].map((address: string) => {
@@ -402,8 +402,8 @@ forking(33763885, async () => {
     it("mint vToken vUSDT", async () => {
       const vUSDTBalance = await USDT.balanceOf(vUSDT.address);
       const usdtHolderBalance = await USDT.balanceOf(await usdtHolder.getAddress());
-      await USDT.connect(usdtHolder).approve(vUSDT.address, 2000);
-      await vUSDT.connect(usdtHolder).mint(2000);
+      await USDT.connect(usdtHolder).approve(vUSDT.address, parseUnits("1", 18));
+      await vUSDT.connect(usdtHolder).mint(parseUnits("1", 18));
       const newvUSDTBalance = await USDT.balanceOf(vUSDT.address);
       const newUsdtHolderBalance = await USDT.balanceOf(await usdtHolder.getAddress());
 
@@ -412,33 +412,37 @@ forking(33763885, async () => {
     });
 
     it("redeem vToken", async () => {
-      await USDT.connect(usdtHolder).approve(vUSDT.address, 2000);
-      await expect(vUSDT.connect(usdtHolder).mint(2000)).to.emit(vUSDT, "Mint");
+      await USDT.connect(usdtHolder).approve(vUSDT.address, parseUnits("1", 18));
+      await expect(vUSDT.connect(usdtHolder).mint(parseUnits("1", 18))).to.emit(vUSDT, "Mint");
 
       const vUSDTUserBal = await vUSDT.connect(usdtHolder).balanceOf(await usdtHolder.getAddress());
-      await expect(vUSDT.connect(usdtHolder).redeem(2000)).to.emit(vUSDT, "Redeem");
+      await expect(vUSDT.connect(usdtHolder).redeem(parseUnits("1", 8))).to.emit(vUSDT, "Redeem");
       const newVUSDTUserBal = await vUSDT.connect(usdtHolder).balanceOf(await usdtHolder.getAddress());
 
-      expect(newVUSDTUserBal).to.equal(vUSDTUserBal.sub(2000));
+      expect(newVUSDTUserBal).to.equal(vUSDTUserBal.sub(parseUnits("1", 8)));
     });
 
     it("borrow vToken", async () => {
       const usdtUserBal = await USDT.balanceOf(await usdtHolder.getAddress());
 
-      await expect(vUSDT.connect(usdtHolder).borrow(1000)).to.emit(vUSDT, "Borrow");
+      await expect(vUSDT.connect(usdtHolder).borrow(parseUnits("1", 18))).to.emit(vUSDT, "Borrow");
 
-      expect((await USDT.balanceOf(await usdtHolder.getAddress())).toString()).to.equal(usdtUserBal.add(1000));
+      expect((await USDT.balanceOf(await usdtHolder.getAddress())).toString()).to.equal(
+        usdtUserBal.add(parseUnits("1", 18)),
+      );
     });
 
     it("Repay vToken", async () => {
-      await USDT.connect(usdtHolder).approve(vUSDT.address, 2000);
+      await USDT.connect(usdtHolder).approve(vUSDT.address, parseUnits("1", 18));
+      await expect(vUSDT.connect(usdtHolder).mint(parseUnits("1", 18))).to.emit(vUSDT, "Mint");
 
       const usdtUserBal = await USDT.balanceOf(await usdtHolder.getAddress());
-      await vUSDT.connect(usdtHolder).borrow(1000);
+      await vUSDT.connect(usdtHolder).borrow(parseUnits("0.5", 18));
 
       expect((await USDT.balanceOf(await usdtHolder.getAddress())).toString()).to.greaterThan(usdtUserBal);
 
-      await vUSDT.connect(usdtHolder).repayBorrow(1000);
+      await USDT.connect(usdtHolder).approve(vUSDT.address, parseUnits("0.5", 18));
+      await vUSDT.connect(usdtHolder).repayBorrow(parseUnits("0.5", 18));
 
       const balanceAfterRepay = await USDT.balanceOf(await usdtHolder.getAddress());
       expect(balanceAfterRepay).to.equal(usdtUserBal);
